# services/learn-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing hiragana and katakana characters",
        "title": "JapaneseStudent Characters API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "shelyahin.mihail@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v3",
    "paths": {
...
        "/api/v1/words": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a mixed list of old and new words for the authenticated user. Requires authentication.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dictionary"
                ],
                "summary": "Get word list",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Number of new words (10-40), default: 20",
                        "name": "newCount",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of old words (10-40), default: 20",
                        "name": "oldCount",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Locale: en, ru, or de, default: en",
                        "name": "locale",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of words",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.WordResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - authentication required",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/words/results": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Submit word learning results with period values. Requires authentication.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dictionary"
                ],
                "summary": "Submit word learning results",
                "parameters": [
                    {
                        "description": "Word results",
                        "name": "results",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.SubmitWordResultsRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad request - invalid request body or validation error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - authentication required",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.SubmitWordResultsRequest": {
            "type": "object",
            "properties": {
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.WordResult"
                    }
                }
            }
        },
...
        "models.WordResponse": {
            "type": "object",
            "properties": {
                "easyPeriod": {
                    "type": "integer"
                },
                "example": {
                    "type": "string"
                },
                "exampleTranslation": {
                    "description": "Locale-specific example translation",
                    "type": "string"
                },
                "extraHardPeriod": {
                    "type": "integer"
                },
                "hardPeriod": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "normalPeriod": {
                    "type": "integer"
                },
                "phoneticClues": {
                    "type": "string"
                },
                "translation": {
                    "description": "Locale-specific word translation",
                    "type": "string"
                },
                "word": {
                    "type": "string"
                }
            }
        },
        "models.WordResult": {
            "type": "object",
            "properties": {
                "period": {
                    "description": "Days (1-30)",
                    "type": "integer"
                },
                "wordId": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}